% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-repo-info.R
\name{gh_get_repo_status}
\alias{gh_get_repo_status}
\title{gh_url}
\usage{
gh_get_repo_status(repos)
}
\arguments{
\item{repos}{list of repositories}
}
\description{
Function that creates csv with status repositories.
}
