% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-repo-info.R
\name{gh_runs}
\alias{gh_runs}
\title{gh_runs}
\usage{
gh_runs(owner, repo, workflow_id, ...)
}
\arguments{
\item{owner}{repository's owner}

\item{repo}{repository's name}

\item{workflow_id}{workflow id}

\item{...}{others paramaters}
}
\description{
Function to create a memorized copy of the repository wokflow id.
}
