% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh-repo-info.R
\name{gh_workflows}
\alias{gh_workflows}
\title{gh_workflows}
\usage{
gh_workflows(owner, repo, ...)
}
\arguments{
\item{owner}{repository's owner}

\item{repo}{repository's name}

\item{...}{others paramaters}
}
\description{
Function to create a memorized copy of the repository cache.
}
